Refactored: moved replicate function from v2 in replicate file
